Read files with UTF-8 encoding instead of platform encoding
Previously several parts of the code use a `FileReader` for reading files, so the default platform encoding was used. Now UTF-8 is used regardless of platform.
Also the `IOUtils` are dropped and the parts of the code that used the `IOUtils` now use try-with blocks.

This should fix #76 (https://github.com/Gubaer/josm-scripting-plugin/issues/76).